Fixed the calculation of the total bytes that a ConcurrentMMFQueue will occupy

+++ Libraries/Foat/Collections/Generic/MMF/ConcurrentMMFQueue.cs
@@ -52,7 +52,7 @@
 
             this.BytesPerItem = bytesPerItem;
             this.ItemCapacity = itemCapacity;
-            this.TotalBytes = itemCapacity * bytesPerItem;
+            this.TotalBytes = (long)itemCapacity * (long)bytesPerItem;
 
             this.ItemCacheCapacity = itemCacheCapacity;
